separated logistics login page components
